chore(api): swap spoke/hub conversion of PaasConfig
